Создал отдельный класс "фабрику" для генерации новых вопросов. Ответственности вынесены по отдельным файлам
