+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Mir on 18.03.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel { 
+    let title: String
+    let text: String
+    let buttonText: String
+}
