+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Mir on 18.03.2023.
+//
+
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
